MTL loss function functioning

+++ data/load_radchar.py
@@ -14,8 +14,8 @@ class RadCharDataset(Dataset):
         return (len(self.data))
 
     def __getitem__(self, index):
-        label = torch.from_numpy(np.asarray(self.labels[index].tolist())[1:6])
-        data = torch.from_numpy(np.concatenate((np.real(self.data[index]).reshape(1,512), np.imag(self.data[index]).reshape(1,512)), axis=0))
+        label = torch.from_numpy(np.asarray(self.labels[index].tolist())[1:6].astype(np.float32))
+        data = torch.from_numpy(np.concatenate((np.real(self.data[index]).astype(np.float32).reshape(1,512), np.imag(self.data[index]).astype(np.float32).reshape(1,512)), axis=0))
         sample = {'data':data ,'label':label}
 
         return sample
